Add Inter font and responsive text utility in app styles
Introduce the Inter font family with a semi-bold style for enhanced typography in the application. This change includes the addition of the Inter_18pt-SemiBold.ttf font file and its registration in the pubspec.yaml file.

Furthermore, a new utility class `AppStyles` has been implemented in the `app_styles.dart` file within the `lib/core/utils` directory. This class provides a method to define text styles responsively based on the screen width, ensuring consistent and readable text across various device sizes. The `semiBold58` text style is an example of how to use the Inter font within the application. Additionally, the `getResponsiveText` method calculates a responsive font size based on a base size and screen dimensions, with clamping to maintain readability. The `getScaleFactor` method is used internally to determine the scaling factor relative to a standard screen width of 400 units.