Refactor HomeView to use Stack for bottom navigation bar positioning
Remove deprecated CrystalNavigationBar and update HomeView to use a Stack for positioning the bottom navigation bar. This change enhances the layout flexibility and prepares the ground for future enhancements to the navigation bar. Additionally, modify CustomBottomIcon to accept IconData instead of Icon for better flexibility and consistency in icon usage.