Refactor custom bottom modal sheet and logout button widgets
Adjusted padding and divider properties in the `CustomBottomModalSheet` widget to improve visual consistency. Specifically, reduced horizontal padding from 12.0 to 11.0, decreased the divider thickness from 4 to 3, and changed the divider color to a slightly darker shade of grey. Additionally, added a fixed height to the SizedBox within the modal sheet for better spacing.

In the `CustomLogOutButton` widget, replaced the `Container` with a `SizedBox` to simplify the widget tree. Increased the button height from 50 to 54 for better touch target size. Removed the redundant `BorderRadius` and `Border` properties from the `Container` and moved them to the `ElevatedButton`'s `style` property for cleaner code.